src/ThreatDetector/utils/__init__.py
import sys 
import yaml
from src.ThreatDetector.exception import CustomException
from src.ThreatDetector.constants import *



def read_yaml(file_path: str) -> dict:
    """
    Reads a YAML file and returns its content as a dictionary.

    Args:
        file_path (str): The path to the YAML file.

    Returns:
        dict: The content of the YAML file as a dictionary.

    Raises:
        CustomException: If an error occurs while reading the YAML file.
    """
    try:
        with open(file_path, "rb") as yaml_file:
            return yaml.safe_load(yaml_file)
    except Exception as e:
        raise CustomException(e, sys)
    
    