requirements.txt
torch
torchvision
pyyaml
from-root
python-dotenv
gdown